requirements.txt
discord.py[voice] >= 2.3.0
PyNaCl >= 1.5.0
openai-whisper >= 20230314
python-dotenv >= 1.0.0
aiosqlite >= 0.19.0
python-Levenshtein >= 0.21.0
pydub >= 0.25.0
ffmpeg4discord >= 0.1.0

